return productID when create and add validation when delate item

+++ src/products/products.controller.ts
@@ -1,13 +1,13 @@
-import { Body, Controller, Delete, Get, HttpCode, HttpException, Param, Patch, Post, Put, Req, UseInterceptors, UsePipes, ValidationPipe } from '@nestjs/common';
+import { Body, Controller, Delete, Get, HttpCode, HttpException, HttpStatus, Param, Patch, Post, Put, Req, UseInterceptors, UsePipes, ValidationPipe } from '@nestjs/common';
 import { Request, Response } from 'express';
 import { ProductsService } from './products.service';
 import { UpdateProductDto } from './dto/update-product.dto';
 import { LoggingInterceptor } from 'src/common/interceptor/logging.interceptor';
 import { CreateProductDto } from './dto/create-product.dto';
-//import { Product } from './schemas/product.schema';
-import mongoose from 'mongoose';
 import { InventoryDto } from './dto/inventory.dto';
 import { Product } from './schemas/product.schema';
+import mongoose from 'mongoose';
+import { IsNumber } from 'class-validator';
 
 @Controller('products')
 @UseInterceptors(LoggingInterceptor)
@@ -21,8 +21,6 @@ export class ProductsController {
 
     @Get(':productID')
     async findOne(@Param('productID') productID: number) {
-        //const isValid = mongoose.Types.ObjectId.isValid(productID);
-        //if (!isValid) throw new HttpException('Product not found', 404);
         const findProduct = await this.productsService.findOne(productID);
         if (!findProduct) throw new HttpException('Product not found', 404);
         return findProduct;
@@ -30,17 +28,17 @@ export class ProductsController {
 
 
     @Post()
-    @HttpCode(204)
+    @HttpCode(HttpStatus.CREATED)
     @UsePipes(new ValidationPipe)
-    create(@Body() createProductDto: CreateProductDto) {
-        console.log('create');
-        this.productsService.create(createProductDto);
+    async create(@Body() createProductDto: CreateProductDto): Promise<number> {
+        const product = await this.productsService.create(createProductDto);
+        return product.product.productID;
     }
 
     @Delete(':productID')
     async deleteProduct(@Param('productID') productID: number) {
-        //const isValid = mongoose.Types.ObjectId.isValid(id);
-        //if (!isValid) throw new HttpException('Invalid ID', 400);
+        const isValid = this.productsService.isNumber(productID);
+        if (!isValid) throw new HttpException('Invalid ID', 400);
         const deleteProduct = await this.productsService.deleteProduct(productID);
         if (!deleteProduct) throw new HttpException('Product Not Found', 404);
         return;
@@ -54,7 +52,6 @@ export class ProductsController {
     ) {
         //const isValid = mongoose.Types.ObjectId.isValid(id);
         //if (!isValid) throw new HttpException('Invalid ID', 400);
-        //(await this.findOne(productID)).
         const updateProduct = await this.productsService.updateProduct(productID, updateProductDto);
         if (!updateProduct) throw new HttpException('Product Not Found', 404);
         return updateProduct;
